readme updates
"heater" is the app we want to optimize/provide/keep working, right? In our previous conversations, I'd thought that the "heater" was the thing that was **stressing** the web app (but that's the siege engine) that we want to optimize/keep working.

I don't know if "This project stores terraform state separately for each layer. " is necessary. I wonder if it is either self-evident for some people and beyond the necessary scope for others. Just knowing that it saves state in S3/Spaces might be enough.

Is the seige app smart enough that I could point it at a Discourse instance and it'd load multiple pages? It might help to explain more about whether it's just loading the root page over and over as fast as possible or if it's doing something more sophisticated.